power-control-host/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "power-control-host"
version = "0.1.0"
description = "Reliability power supply control host skeleton."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "pyvisa>=1.14.1",
  "pyvisa-py>=0.7.2",
  "pyserial>=3.5",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
power-control-host = "power_control_host.ui.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

